Please read details about changes
Based on the census documentation we can not do year over year comparisons using the ACS5 year data. We cannot have overlapping years in the 5 year average calcs.  Thus, I created a loop that brought in 2014 and 2019 data into a single data frame.  I then read in the csv file from Eric that had MSA name. and created a merged_census_df.  I had to change the datatype in the census_df to be integer instead of object before I could  with the df from jack's file.  We have approx 148 zipcodes that did not map.  We also have some odd 66666 data in the 2019 dataframe that we will need to drop to clean the dataframe.  Sending these comments to you in slack and the excel file that I exported the combined_census_dataframe.