Updates to the loop of the main datafraame
1)Updates to the loop of the main dataframe. 2) Cleans data for both 2014 and 2019. 3) merges MSA data, changes value type for columns. 4)First US Censdus dataframe now just called df.
NOTE TO ERIC:ERIC,the main data frame is called df.  Also, the columns are the original names is _2014 and _2019 as a suffix.  Eric from 21 forward where you are calculating the measures, etc., you will now need to refer to the column year specific names.  That is where the code is breaking.